refactor config, move all conf related code to a separate package
