+++ app/config/testdata/file.txt
@@ -1,0 +1,1 @@
+Not Yaml
