+++ app/proc/processor.go
@@ -4,14 +4,13 @@ package proc
 
 import (
 	"context"
-	"regexp"
 	"time"
 
 	log "github.com/go-pkgz/lgr"
 	"github.com/go-pkgz/syncs"
 
+	"github.com/umputun/feed-master/app/config"
 	"github.com/umputun/feed-master/app/feed"
-	"github.com/umputun/feed-master/app/youtube"
 )
 
 // TelegramNotif is interface to send messages to telegram
@@ -26,71 +25,16 @@ type TwitterNotif interface {
 
 // Processor is a feed reader and store writer
 type Processor struct {
-	Conf          *Conf
+	Conf          *config.Conf
 	Store         *BoltDB
 	TelegramNotif TelegramNotif
 	TwitterNotif  TwitterNotif
 }
 
-// Conf for feeds config yml
-type Conf struct {
-	Feeds  map[string]Feed `yaml:"feeds"`
-	System struct {
-		UpdateInterval time.Duration `yaml:"update"`
-		MaxItems       int           `yaml:"max_per_feed"`
-		MaxTotal       int           `yaml:"max_total"`
-		MaxKeepInDB    int           `yaml:"max_keep"`
-		Concurrent     int           `yaml:"concurrent"`
-		BaseURL        string        `yaml:"base_url"`
-	} `yaml:"system"`
-
-	YouTube struct {
-		DlTemplate      string             `yaml:"dl_template"`
-		BaseChanURL     string             `yaml:"base_chan_url"`
-		BasePlaylistURL string             `yaml:"base_playlist_url"`
-		Channels        []youtube.FeedInfo `yaml:"channels"`
-		BaseURL         string             `yaml:"base_url"`
-		UpdateInterval  time.Duration      `yaml:"update"`
-		MaxItems        int                `yaml:"max_per_channel"`
-		FilesLocation   string             `yaml:"files_location"`
-		RSSLocation     string             `yaml:"rss_location"`
-	} `yaml:"youtube"`
-}
-
-// Source defines config section for source
-type Source struct {
-	Name string `yaml:"name"`
-	URL  string `yaml:"url"`
-}
-
-// Feed defines config section for a feed~
-type Feed struct {
-	Title           string   `yaml:"title"`
-	Description     string   `yaml:"description"`
-	Link            string   `yaml:"link"`
-	Image           string   `yaml:"image"`
-	Language        string   `yaml:"language"`
-	TelegramChannel string   `yaml:"telegram_channel"`
-	Filter          Filter   `yaml:"filter"`
-	Sources         []Source `yaml:"sources"`
-	ExtendDateTitle string   `yaml:"ext_date"`
-}
-
-// Filter defines feed section for a feed filter~
-type Filter struct {
-	Title string `yaml:"title"`
-}
-
-// YTChannel defines youtube channel config
-type YTChannel struct {
-	ID   string
-	Name string
-}
-
 // Do activate loop of goroutine for each feed, concurrency limited by p.Conf.Concurrent
 func (p *Processor) Do() {
 	log.Printf("[INFO] activate processor, feeds=%d, %+v", len(p.Conf.Feeds), p.Conf)
-	p.setDefaults()
+	config.SetDefaults(p.Conf)
 
 	for {
 		swg := syncs.NewSizedGroup(p.Conf.System.Concurrent, syncs.Preemptive)
@@ -116,7 +60,7 @@ func (p *Processor) Do() {
 	}
 }
 
-func (p *Processor) feed(name, url, telegramChannel string, max int, filter Filter) {
+func (p *Processor) feed(name, url, telegramChannel string, max int, filter config.Filter) {
 	rss, err := feed.Parse(url)
 	if err != nil {
 		log.Printf("[WARN] failed to parse %s, %v", url, err)
@@ -135,7 +79,7 @@ func (p *Processor) feed(name, url, telegramChannel string, max int, filter Filt
 			continue
 		}
 
-		skip, err := filter.skip(item)
+		skip, err := filter.Skip(item)
 		if err != nil {
 			log.Printf("[WARN] failed to filter %s (%s) to %s, save as is, %v", item.GUID, item.PubDate, name, err)
 		}
@@ -164,36 +108,4 @@ func (p *Processor) feed(name, url, telegramChannel string, max int, filter Filt
 			log.Printf("[WARN] failed send twitter message, url=%s, %v", item.Enclosure.URL, err)
 		}
 	}
-}
-
-func (p *Processor) setDefaults() {
-	if p.Conf.System.Concurrent == 0 {
-		p.Conf.System.Concurrent = 8
-	}
-	if p.Conf.System.MaxItems == 0 {
-		p.Conf.System.MaxItems = 5
-	}
-	if p.Conf.System.MaxTotal == 0 {
-		p.Conf.System.MaxTotal = 100
-	}
-	if p.Conf.System.MaxKeepInDB == 0 {
-		p.Conf.System.MaxKeepInDB = 5000
-	}
-	if p.Conf.System.UpdateInterval == 0 {
-		p.Conf.System.UpdateInterval = time.Minute * 5
-	}
-}
-
-func (filter *Filter) skip(item feed.Item) (bool, error) {
-	if filter.Title != "" {
-		matched, err := regexp.MatchString(filter.Title, item.Title)
-		if err != nil {
-			return matched, err
-		}
-		if matched {
-			return true, err
-		}
-	}
-
-	return false, nil
 }
